Refactor: decompose ServiceAccountDetail into sibling files
ServiceAccountDetail.tsx was 1240 LOC: 13 useStates, 9 mutations, two
near-identical "newly-created secret" banners, and three cards worth
of render inline. Extract each concern so touching one stops forcing
a scroll through the others. Pure extraction — no behavior change.

- src/components/ui/secret-banner.tsx: generic SecretBanner with
  internal copy-to-clipboard state keyed by secret label. Replaces
  the two near-duplicate banners for new credentials and new API
  keys.
- useServiceAccountMutations.ts: wraps all nine mutations (add /
  update-role / remove org membership; create / revoke / rotate
  credentials; create / revoke / rotate API keys) plus their
  queryClient invalidations. Error handling via extractApiErrorDetail
  preserved exactly; per-call onSuccess stays with each caller so
  local form resets still fire.
- OrgMembershipsCard.tsx: org-membership UI. Owns its own showAddOrg
  / newOrgSlug / newRoleSlug state. Requests destructive confirms
  via an onConfirmRemove callback; the ConfirmDialog stays in the
  parent.
- ClientCredentialsSection.tsx: client-credentials UI. Owns its
  five local form fields (name, description, scopes, expiresDays,
  showCreate*). Uses SecretBanner for the newly-created state.
- ApiKeysSection.tsx: API-keys UI, same shape as credentials.
  Owns newKeyName + showCreateKey locally.

The 13 useState concentration in the parent naturally redistributes
across the three sub-components; useReducer isn't needed. Parent
retains three useStates that genuinely cross section boundaries:
the confirm discriminated-union and the two newly-created-secret
holders (the rotate ConfirmDialog lives in the parent, so its
success handler needs to set the banner state consumed by the
child). All five ConfirmDialogs stay in the parent.

Parent is now 329 LOC.
